Add class AlternativeHedonicPrice to LT. Add floor_area field to PopulationPerPlanningArea function. Add hedonic_price to alternative class. Refactor code. Update the formula used for the screening model.

+++ dev/Basic/long/database/entity/PopulationPerPlanningArea.cpp
@@ -52,7 +52,7 @@ int PopulationPerPlanningArea::getPopulation() const
 	return population;
 }
 
-int PopulationPerPlanningArea::getFloorArea() const
+double PopulationPerPlanningArea::getFloorArea() const
 {
 	return floorArea;
 }
